finish

+++ lab4/server.py
@@ -12,7 +12,7 @@ def new_client(clientsocket, addr):
 	def sql_unsubscribe_board(conn, c, uid, msg_list):
 		sql_return = c.execute('select * from SUB_BOARD where Board_name = ? and Subscriber_id = ?', (msg_list[2], uid)).fetchone()
 		if sql_return != None:
-			c.execute('delete from SUB_BOARD where Board_name = ?', (msg_list[2],))
+			c.execute('delete from SUB_BOARD where Board_name = ? and Subscriber_id = ?', (msg_list[2], uid))
 			conn.commit()
 			msg_suc = 'Unsubscribe successfully\r\n'
 		else:
@@ -23,7 +23,7 @@ def new_client(clientsocket, addr):
 	def sql_unsubscribe_author(conn, c, uid, msg_list):
 		sql_return = c.execute('select * from SUB_AUTHOR where Author_name = ? and Subscriber_id = ?', (msg_list[2], uid)).fetchone()
 		if sql_return != None:
-			c.execute('delete from SUB_AUTHOR where Author_name = ?', (msg_list[2],))
+			c.execute('delete from SUB_AUTHOR where Author_name = ? and Subscriber_id = ?', (msg_list[2], uid))
 			conn.commit()
 			msg_suc = 'Unsubscribe successfully\r\n'
 		else:
